# Conflicts:
#	app/src/main/java/com/example/todo_list_application/di/ApplicationModule.kt
#	data/src/main/java/com/example/data/datasource/db/DBTaskDataSource.kt
#	ui/src/main/java/com/example/ui/listtask/viewmodel/TaskListViewModel.kt

+++ data/src/androidTest/java/com/example/data/datasource/db/dao/TaskDaoTest.kt
@@ -127,7 +127,7 @@ class TaskDaoTest {
         Assert.assertEquals(updatedTask!!.isPinned, false)
     }
 
-    private fun getTask(id: Int, title: String, description: String, isComplete: Boolean): Task = Task(
+    private fun getTask(id: Long, title: String, description: String, isComplete: Boolean): Task = Task(
         id, title, description, isComplete, Date(), Date()
     )
